requirements.txt
# OCR lent (image → texte)
pytesseract==0.3.10

# Gestion des images
Pillow==12.1.0

# PDF → images pour OCR
pdf2image==1.16.3

# PDF → texte natif rapide
pdfplumber==0.11.4

# HTTP pour callbacks / communication
requests==2.31.0
